Dialog for proximity change Demo data changed to a single injection at the app beginning, so that the data can change.
