+++ composeApp/src/commonMain/kotlin/data/io/user/NetworkItemIO.kt
@@ -26,6 +26,7 @@ data class NetworkItemIO @OptIn(ExperimentalUuidApi::class) constructor(
     val publicId: String = Uuid.random().toString(),
 
     /** current user's public id, generated by our BE - not associated with Firebase */
+    @ColumnInfo("user_public_id")
     val userPublicId: String? = null,
 
     /** Matrix public user identifier */
