+++ composeApp/src/commonMain/kotlin/ui/network/add_new/NetworkAddNewRepository.kt
@@ -5,6 +5,10 @@ import data.io.base.BaseResponse
 import data.io.social.network.request.CircleRequestResponse
 import data.io.social.network.request.CirclingRequest
 import data.io.user.NetworkItemIO
+import database.dao.ConversationRoomDao
+import database.dao.NetworkItemDao
+import dev.gitlive.firebase.Firebase
+import dev.gitlive.firebase.auth.auth
 import io.ktor.client.HttpClient
 import io.ktor.client.request.post
 import io.ktor.client.request.setBody
@@ -14,7 +18,11 @@ import kotlinx.coroutines.withContext
 import ui.login.safeRequest
 
 /** Class for calling APIs and remote work in general */
-class NetworkAddNewRepository(private val httpClient: HttpClient) {
+class NetworkAddNewRepository(
+    private val httpClient: HttpClient,
+    private val conversationRoomDao: ConversationRoomDao,
+    private val networkItemDao: NetworkItemDao
+) {
 
     /** Acts upon a circling request */
     suspend fun includeNewUser(action: CirclingRequest): BaseResponse<CircleRequestResponse> {
@@ -31,10 +39,34 @@ class NetworkAddNewRepository(private val httpClient: HttpClient) {
     }
 
     /** Returns list of recommended users from each proximity category */
-    suspend fun getUserRecommendations(): BaseResponse<Map<NetworkProximityCategory, List<NetworkItemIO>>> {
+    suspend fun getUserRecommendations(
+        takeCount: Int,
+        excludeId: String?
+    ): BaseResponse<Map<NetworkProximityCategory, List<NetworkItemIO>>> {
         return withContext(Dispatchers.IO) {
-            // TODO local Room database required here
-            BaseResponse.Error()
+            BaseResponse.Success(
+                NetworkProximityCategory.entries.associateWith { category ->
+                    (conversationRoomDao.getByProximity(
+                        ownerPublicId = Firebase.auth.currentUser?.uid,
+                        proximityMin = category.range.start,
+                        proximityMax = category.range.endInclusive,
+                        count = takeCount,
+                        excludeId = excludeId
+                    ).map { conversation ->
+                        NetworkItemIO(
+                            name = conversation.summary?.alias,
+                            photoUrl = conversation.summary?.avatarUrl,
+                            tag = conversation.summary?.tag
+                        )
+                    } + networkItemDao.getByProximity(
+                        ownerPublicId = Firebase.auth.currentUser?.uid,
+                        proximityMin = category.range.start,
+                        proximityMax = category.range.endInclusive,
+                        count = takeCount,
+                        excludeId = excludeId
+                    )).shuffled().take(takeCount)
+                }
+            )
         }
     }
 }
