+++ composeApp/src/commonMain/kotlin/ui/network/list/NetworkListScreen.kt
@@ -3,81 +3,83 @@ package ui.network.list
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.expandVertically
 import androidx.compose.animation.fadeIn
-import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.gestures.Orientation
+import androidx.compose.foundation.gestures.detectTapGestures
+import androidx.compose.foundation.gestures.draggable
+import androidx.compose.foundation.gestures.rememberDraggableState
+import androidx.compose.foundation.gestures.scrollBy
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.GridItemSpan
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.lazy.grid.rememberLazyGridState
 import androidx.compose.material.Divider
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.windowsizeclass.WindowWidthSizeClass
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.unit.dp
 import androidx.paging.LoadState
 import app.cash.paging.compose.collectAsLazyPagingItems
 import augmy.composeapp.generated.resources.Res
 import augmy.composeapp.generated.resources.network_list_empty_action
 import augmy.composeapp.generated.resources.network_list_empty_title
+import augmy.interactive.shared.ext.scalingClickable
+import augmy.interactive.shared.ui.base.LocalDeviceType
 import augmy.interactive.shared.ui.base.LocalNavController
 import augmy.interactive.shared.ui.theme.LocalTheme
 import base.navigation.NavigationArguments
 import base.utils.getOrNull
 import collectResult
 import components.EmptyLayout
-import components.OptionsLayout
-import components.OptionsLayoutAction
 import components.network.NetworkItemRow
 import components.pull_refresh.RefreshableContent
 import components.pull_refresh.RefreshableViewModel.Companion.requestData
 import data.NetworkProximityCategory
+import data.io.user.NetworkItemIO
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import org.jetbrains.compose.resources.stringResource
 import org.koin.compose.viewmodel.koinViewModel
+import ui.network.RefreshHandler
+import ui.network.components.SocialItemActions
+import ui.network.profile.UserProfileLauncher
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid
 
 /** Screen containing current user's network and offers its management */
-@OptIn(ExperimentalUuidApi::class, ExperimentalFoundationApi::class)
+@OptIn(ExperimentalUuidApi::class, ExperimentalMaterial3Api::class)
 @Composable
-fun  NetworkListContent(
+fun NetworkListContent(
     openAddNewModal: () -> Unit,
+    refreshHandler: RefreshHandler,
     viewModel: NetworkListViewModel = koinViewModel()
 ) {
-    val networkItems = viewModel.requests.collectAsLazyPagingItems()
+    val networkItems = viewModel.networkItems.collectAsLazyPagingItems()
     val isRefreshing = viewModel.isRefreshing.collectAsState()
     val customColors = viewModel.customColors.collectAsState(initial = mapOf())
 
+    val coroutineScope = rememberCoroutineScope()
+    val listState = rememberLazyGridState()
     val navController = LocalNavController.current
     val isLoadingInitialPage = networkItems.loadState.refresh is LoadState.Loading
 
     val checkedItems = remember { mutableStateListOf<String?>() }
     val selectedItem = remember { mutableStateOf<String?>(null) }
-    val coroutineScope = rememberCoroutineScope()
-
-    val onAction: (OptionsLayoutAction) -> Unit = { action ->
-        when(action) {
-            OptionsLayoutAction.Invite -> {
-                // TODO
-            }
-            OptionsLayoutAction.DeselectAll -> checkedItems.clear()
-            OptionsLayoutAction.SelectAll -> {
-                coroutineScope.launch(Dispatchers.Default) {
-                    checkedItems.addAll(
-                        checkedItems.toMutableSet().apply {
-                            addAll(networkItems.itemSnapshotList.items.map { it.userPublicId })
-                        }
-                    )
-                }
-            }
-            else -> {}
-        }
+    val selectedUser = remember {
+        mutableStateOf<NetworkItemIO?>(null)
     }
-
 
     navController?.collectResult(
         key = NavigationArguments.NETWORK_NEW_SUCCESS,
@@ -87,6 +89,22 @@ fun  NetworkListContent(
         }
     )
 
+    if(selectedUser.value != null) {
+        UserProfileLauncher(
+            userProfile = selectedUser.value,
+            onDismissRequest = {
+                selectedUser.value = null
+            }
+        )
+    }
+
+    LaunchedEffect(Unit) {
+        refreshHandler.addListener {
+            viewModel.requestData(isSpecial = true, isPullRefresh = true)
+            networkItems.refresh()
+        }
+    }
+
     RefreshableContent(
         onRefresh = {
             viewModel.requestData(isSpecial = true, isPullRefresh = true)
@@ -94,18 +112,29 @@ fun  NetworkListContent(
         },
         isRefreshing = isRefreshing.value
     ) {
-        LazyColumn(
-            modifier = Modifier.fillMaxSize(),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            stickyHeader {
-                OptionsLayout(
-                    modifier = Modifier.animateItem(),
-                    show = checkedItems.size > 0,
-                    onClick = onAction
+        LazyVerticalGrid(
+            modifier = Modifier
+                .pointerInput(Unit) {
+                    detectTapGestures(onTap = {
+                        selectedItem.value = null
+                    })
+                }
+                .draggable(
+                    orientation = Orientation.Vertical,
+                    state = rememberDraggableState { delta ->
+                        coroutineScope.launch {
+                            listState.scrollBy(-delta)
+                        }
+                    }
                 )
-            }
-            item {
+                .fillMaxSize(),
+            columns = GridCells.Fixed(
+                if(LocalDeviceType.current == WindowWidthSizeClass.Compact) 1 else 2
+            ),
+            state = listState,
+            verticalArrangement = Arrangement.spacedBy(LocalTheme.current.shapes.betweenItemsSpace)
+        ) {
+            item(span = { GridItemSpan(maxLineSpan) }) {
                 AnimatedVisibility(
                     enter = expandVertically() + fadeIn(),
                     visible = networkItems.itemCount == 0 && !isLoadingInitialPage
@@ -124,6 +153,32 @@ fun  NetworkListContent(
                 networkItems.getOrNull(index).let { data ->
                     NetworkItemRow(
                         modifier = Modifier
+                            .scalingClickable(
+                                hoverEnabled = selectedItem.value != data?.userPublicId,
+                                scaleInto = .9f,
+                                onTap = {
+                                    selectedItem.value = if(selectedItem.value != data?.userPublicId) {
+                                        data?.userPublicId
+                                    }else null
+                                },
+                                onLongPress = {
+                                    selectedItem.value = data?.userPublicId
+                                }
+                            )
+                            .then(
+                                (if(selectedItem.value != null && selectedItem.value == data?.userPublicId) {
+                                    Modifier
+                                        .background(
+                                            color = LocalTheme.current.colors.backgroundLight,
+                                            shape = LocalTheme.current.shapes.rectangularActionShape
+                                        )
+                                        .border(
+                                            width = 2.dp,
+                                            color = LocalTheme.current.colors.backgroundDark,
+                                            shape = LocalTheme.current.shapes.rectangularActionShape
+                                        )
+                                }else Modifier)
+                            )
                             .fillMaxWidth()
                             .animateItem(),
                         isChecked = if(checkedItems.size > 0) {
@@ -135,7 +190,33 @@ fun  NetworkListContent(
                         }.let {
                             customColors.value[it] ?: it?.color
                         },
+                        onAvatarClick = {
+                            coroutineScope.launch(Dispatchers.Default) {
+                                selectedUser.value = data
+                            }
+                        },
                         isSelected = selectedItem.value == data?.userPublicId,
+                        actions = {
+                            if(data != null) {
+                                SocialItemActions(
+                                    key = data.userPublicId,
+                                    requestProximityChange = { proximity ->
+                                        viewModel.requestProximityChange(
+                                            publicId = data.publicId,
+                                            proximity = proximity,
+                                            onOperationDone = {
+                                                if(selectedItem.value == data.userPublicId) {
+                                                    selectedItem.value = null
+                                                }
+                                                networkItems.refresh()
+                                            }
+                                        )
+                                    },
+                                    onInvite = {},
+                                    newItem = data
+                                )
+                            }
+                        }
                         /*onCheckChange = { isLongClick ->
                             when {
                                 checkedItems.contains(data?.userPublicId) -> checkedItems.remove(data?.userPublicId)
